robot_module refactor

+++ code/game/objects/items/stacks/stack.dm
@@ -10,6 +10,10 @@
  */
 /obj/item/stack
 	origin_tech = "materials=1"
+	var/is_cyborg = FALSE // It's TRUE if the module is a cyborg stack.
+	var/datum/robot_energy_storage/source
+	var/energy_type // Which robot_energy_storage to choose. i.e. /datum/robot_energy_storage/metal
+	var/cost = 1 // How much energy from storage it costs
 	var/list/recipes = list() // /datum/stack_recipe
 	var/singular_name
 	var/amount = 1
@@ -50,15 +54,27 @@
 
 /obj/item/stack/examine(mob/user)
 	. = ..()
-	if(in_range(user, src))
+	if(!in_range(user, src))
+		return
+
+	if(is_cyborg)
 		if(singular_name)
-			. += "There are [amount] [singular_name]\s in the stack."
+			. += "There is enough energy for [get_amount()] [singular_name]\s."
 		else
-			. += "There are [amount] [name]\s in the stack."
-		. +="<span class='notice'>Alt-click to take a custom amount.</span>"
+			. += "There is enough energy for [get_amount()]."
+		return
+
+	if(singular_name)
+		. += "There are [amount] [singular_name]\s in the stack."
+	else
+		. += "There are [amount] [name]\s in the stack."
+	. +="<span class='notice'>Alt-click to take a custom amount.</span>"
 
 /obj/item/stack/proc/add(newamount)
-	amount += newamount
+	if(is_cyborg)
+		source.add_charge(newamount * cost)
+	else
+		amount += newamount
 	update_icon()
 
 /obj/item/stack/attack_self(mob/user)
@@ -85,8 +101,10 @@
 	if(!recipes)
 		return
 
-	if(amount <= 0)
+	if(get_amount() <= 0)
 		user << browse(null, "window=stack")
+		if(is_cyborg)
+			to_chat(user, "<span class='warning'>You don't have enough energy to dispense more [name]!</span>")
 		return
 
 	user.set_machine(src) //for correct work of onclose
@@ -96,7 +114,7 @@
 		var/datum/stack_recipe_list/srl = recipe_list[recipes_sublist]
 		recipe_list = srl.recipes
 
-	var/t1 = "Amount Left: [amount]<br>"
+	var/t1 = "Amount Left: [get_amount()]<br>"
 	for(var/i in 1 to recipe_list.len)
 		var/E = recipe_list[i]
 		if(isnull(E))
@@ -112,7 +130,7 @@
 
 		if(istype(E, /datum/stack_recipe))
 			var/datum/stack_recipe/R = E
-			var/max_multiplier = round(amount / R.req_amount)
+			var/max_multiplier = round(get_amount() / R.req_amount)
 			var/title
 			var/can_build = 1
 			can_build = can_build && (max_multiplier > 0)
@@ -152,7 +170,7 @@
 		list_recipes(usr, text2num(href_list["sublist"]))
 
 	if(href_list["make"])
-		if(amount < 1)
+		if(amount < 1 && !is_cyborg)
 			qdel(src) //Never should happen
 
 		var/list/recipes_list = recipes
@@ -165,8 +183,8 @@
 		if(!multiplier || multiplier <= 0 || multiplier > 50) // Href exploit checks
 			multiplier = 1
 
-		if(amount < R.req_amount * multiplier)
-			if(R.req_amount * multiplier>1)
+		if(get_amount() < R.req_amount * multiplier)
+			if(R.req_amount * multiplier > 1)
 				to_chat(usr, "<span class='warning'>You haven't got enough [src] to build \the [R.req_amount * multiplier] [R.title]\s!</span>")
 			else
 				to_chat(usr, "<span class='warning'>You haven't got enough [src] to build \the [R.title]!</span>")
@@ -189,7 +207,7 @@
 			if(!do_after(usr, R.time, target = usr))
 				return 0
 
-		if(amount < R.req_amount * multiplier)
+		if(get_amount() < R.req_amount * multiplier)
 			return
 
 		var/atom/O
@@ -199,6 +217,7 @@
 			O = new R.result_type(usr.drop_location())
 		O.setDir(usr.dir)
 		use(R.req_amount * multiplier)
+		updateUsrDialog()
 
 		R.post_build(src, O)
 
@@ -225,6 +244,8 @@
 /obj/item/stack/use(used, check = TRUE)
 	if(check && zero_amount())
 		return FALSE
+	if(is_cyborg)
+		return source.use_charge(used * cost)
 	if(amount < used)
 		return FALSE
 	amount -= used
@@ -234,6 +255,8 @@
 	return TRUE
 
 /obj/item/stack/proc/get_amount()
+	if(is_cyborg)
+		return round(source.energy / cost)
 	return amount
 
 /obj/item/stack/proc/get_max_amount()
@@ -265,6 +288,8 @@
 		to_chat(user, "<span class='warning'>You can't do that right now!</span>")
 		return
 	if(!in_range(src, user))
+		return
+	if(is_cyborg)
 		return
 	if(!ishuman(usr))
 		return
@@ -299,11 +324,9 @@
 // Returns TRUE if the stack amount is zero.
 // Also qdels the stack gracefully if it is.
 /obj/item/stack/proc/zero_amount()
+	if(is_cyborg)
+		return source.energy < cost
 	if(amount < 1)
-		if(isrobot(loc))
-			var/mob/living/silicon/robot/R = loc
-			if(locate(src) in R.module.modules)
-				R.module.modules -= src
 		if(ismob(loc))
 			var/mob/living/L = loc // At this stage, stack code is so horrible and atrocious, I wouldn't be all surprised ghosts can somehow have stacks. If this happens, then the world deserves to burn.
 			L.unEquip(src, TRUE)
@@ -319,7 +342,10 @@
 	if(QDELETED(S) || QDELETED(src) || S == src) //amusingly this can cause a stack to consume itself, let's not allow that.
 		return FALSE
 	var/transfer = get_amount()
-	transfer = min(transfer, S.max_amount - S.amount)
+	if(S.is_cyborg)
+		transfer = min(transfer, round((S.source.max_energy - S.source.energy) / S.cost))
+	else
+		transfer = min(transfer, S.max_amount - S.amount)
 	if(transfer <= 0)
 		return
 	if(pulledby)
@@ -327,6 +353,7 @@
 	S.copy_evidences(src)
 	S.add(transfer)
 	use(transfer)
+	return transfer
 
 /obj/item/stack/proc/copy_evidences(obj/item/stack/from)
 	blood_DNA			= from.blood_DNA
